Throw an IllegalArgumentException once a ClassCastException occurs.
Closes #11775

+++ model/map/src/main/java/org/keycloak/models/map/authSession/MapRootAuthenticationSessionProvider.java
@@ -145,7 +145,7 @@ public class MapRootAuthenticationSessionProvider implements AuthenticationSessi
 
         DefaultModelCriteria<RootAuthenticationSessionModel> mcb = criteria();
         mcb = mcb.compare(SearchableFields.REALM_ID, Operator.EQ, realm.getId())
-          .compare(SearchableFields.EXPIRATION, Operator.LT, Time.currentTime());
+          .compare(SearchableFields.EXPIRATION, Operator.LT, (long) Time.currentTime());
 
         long deletedCount = tx.delete(withCriteria(mcb));
